outsourced ModelingUtil
