+++ app/lib/objectivemodeler/features/label-editing/LabelUtil.js
@@ -1,4 +1,4 @@
-import {isAny} from '../modeling/util/ModelingUtil';
+import {isAny} from '../../../frequentlyUsedFiles/features/modeling/ModelingUtil.js';
 
 function getLabelAttr(semantic) {
   if (semantic.labelAttribute) {
